// src/lib/shared-models/models/ReactionDeltaDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NCorp.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ReactionDeltaDto
 */
export interface ReactionDeltaDto {
    /**
     * 
     * @type {string}
     * @memberof ReactionDeltaDto
     */
    roomId?: string;
    /**
     * 
     * @type {string}
     * @memberof ReactionDeltaDto
     */
    messageId?: string;
    /**
     * 
     * @type {string}
     * @memberof ReactionDeltaDto
     */
    emoji?: string;
    /**
     * 
     * @type {string}
     * @memberof ReactionDeltaDto
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof ReactionDeltaDto
     */
    op?: string;
    /**
     * 
     * @type {number}
     * @memberof ReactionDeltaDto
     */
    count?: number;
}

/**
 * Check if a given object implements the ReactionDeltaDto interface.
 */
export function instanceOfReactionDeltaDto(value: object): value is ReactionDeltaDto {
    return true;
}

export function ReactionDeltaDtoFromJSON(json: any): ReactionDeltaDto {
    return ReactionDeltaDtoFromJSONTyped(json, false);
}

export function ReactionDeltaDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReactionDeltaDto {
    if (json == null) {
        return json;
    }
    return {
        
        'roomId': json['roomId'] == null ? undefined : json['roomId'],
        'messageId': json['messageId'] == null ? undefined : json['messageId'],
        'emoji': json['emoji'] == null ? undefined : json['emoji'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'op': json['op'] == null ? undefined : json['op'],
        'count': json['count'] == null ? undefined : json['count'],
    };
}

export function ReactionDeltaDtoToJSON(json: any): ReactionDeltaDto {
    return ReactionDeltaDtoToJSONTyped(json, false);
}

export function ReactionDeltaDtoToJSONTyped(value?: ReactionDeltaDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'roomId': value['roomId'],
        'messageId': value['messageId'],
        'emoji': value['emoji'],
        'userId': value['userId'],
        'op': value['op'],
        'count': value['count'],
    };
}